Add migrate:import command
Basic usage is:

migrate:import {file?} {--s|schema=}

where file is an optional argument used to specify the file name to
import and --s/schema is an option used to specify the schema name.

For the moment it only writes out the laravel migration code reflecting
the SQL CREATE statements but it will create a fully usable migration
class ready to be applied.

ALTER statements will be supported as well.